+++ tests/testthat/test-error/test-error.R
@@ -1,0 +1,5 @@
+context("test-error")
+
+test_that("should fail", {
+  expect_equal(1 + 1, 3)
+})