+++ tests/testthat/test-test_dir.R
@@ -33,3 +33,25 @@ test_that('filter_test_scripts() with tricky names', {
     files[1]
   )
 })
+
+
+# errors ------------------------------------------------------------------
+
+test_that("can control if failures generate errors",{
+  test_error <- function(...) {
+    test_dir(test_path("test-error"), reporter = "silent", ...)
+  }
+
+  expect_error(test_error(stop_on_failure = TRUE), "Test failures")
+  expect_error(test_error(stop_on_failure = FALSE), NA)
+})
+
+
+test_that("can control if warnings errors",{
+  test_warning <- function(...) {
+    test_dir(test_path("test-warning"), reporter = "silent", ...)
+  }
+
+  expect_error(test_warning(stop_on_warning = TRUE), "Tests generated warnings")
+  expect_error(test_warning(stop_on_warning = FALSE), NA)
+})