+++ tests/testthat/test-warning/test-warning.R
@@ -1,0 +1,5 @@
+context("test-warning")
+
+test_that("warning emitted", {
+  warning("This is not a test", call. = FALSE)
+})